feat(add-section): add section button to allow adding new sections
The SectionAddBtn component was added to the PageSection component to allow users to add new sections to the page.

The component was also added to the project and styled to appear at the top and bottom of the section when hovered.
The defaultBlankSection object was added to the project to be used when adding new sections.
The addSection function was added to the sections store to allow adding new sections to the store.